feat: Adjust on Suplier entity, add condition do create a suplier based on unique cnpj and code and add conditions to edit and delete suplier based on hasMovement

+++ src/main/java/com/edilson/service/SuplierService.java
@@ -22,6 +22,20 @@ public class SuplierService {
     }
 
     public SuplierEntity createSuplier(SuplierEntity suplierEntity) {
+        var suplier = suplierRepository.find("cnpj", suplierEntity.getCnpj())
+            .firstResultOptional();
+        
+        if (suplier.isPresent()) {
+            throw new IllegalArgumentException("CNPJ já cadastrado");
+        }
+
+        var suplierCode = suplierRepository.find("code", suplierEntity.getCode())
+            .firstResultOptional();
+        
+        if (suplierCode.isPresent()) {
+            throw new IllegalArgumentException("Código já cadastrado");
+        }
+        
         suplierRepository.persist(suplierEntity);
         
         return suplierEntity;
@@ -36,22 +50,37 @@ public class SuplierService {
     public SuplierEntity updateSuplier(Long id, SuplierEntity suplierEntity) {
         var suplier = findById(id);
         
-        
-        suplier.setCode(suplierEntity.getCode());
-        suplier.setLegalName(suplierEntity.getLegalName());
-        suplier.setEmail(suplierEntity.getEmail());
-        suplier.setPhone(suplierEntity.getPhone());
-        suplier.setCnpj(suplierEntity.getCnpj());
-        suplier.setSituation(suplierEntity.getSituation());
-        suplier.setDataBaixa(suplierEntity.getDataBaixa());
-        
-        suplierRepository.persist(suplier);
-        
-        return suplier;
+        if (suplier.isHasMovement()) {
+            if (!suplier.getSituation().equals(suplierEntity.getSituation())) {
+                suplier.setSituation(suplierEntity.getSituation());
+                suplierRepository.persist(suplier);
+                
+                return suplier;
+            } else {
+                throw new IllegalArgumentException("Suplier has movement and cannot be updated");
+                
+            }
+        } else {
+            suplier.setCode(suplierEntity.getCode());
+            suplier.setLegalName(suplierEntity.getLegalName());
+            suplier.setEmail(suplierEntity.getEmail());
+            suplier.setPhone(suplierEntity.getPhone());
+            suplier.setCnpj(suplierEntity.getCnpj());
+            suplier.setSituation(suplierEntity.getSituation());
+            suplier.setDataBaixa(suplierEntity.getDataBaixa());
+            
+            suplierRepository.persist(suplier);
+            
+            return suplier;
+        }
     }
 
     public void deleteSuplier(Long id) {
         var suplier = findById(id);
+
+        if (suplier.isHasMovement()) {
+            throw new IllegalArgumentException("Suplier has movement and cannot be deleted");
+        }
         
         suplierRepository.delete(suplier);
     }
